9 experimental studies from Bem (2011) as described in Bem et al. (2011)

The data set contains Cohen's d effect sizes, standard errors, and labels for 9 experimental studies of precognition from the infamous Bem (2011) as analyzed in his later meta-analysis (Bem et al. 2011) .

Usage

Bem2011

Format

A data.frame with 3 columns and 9 observations:

d

Cohen's d effect size.

se

Standard error of d.

study

Study label.

References

Bem DJ (2011). “Feeling the future: Experimental evidence for anomalous retroactive influences on cognition and affect.” Journal of Personality and Social Psychology, 100(3), 407–425. doi:10.1037/a0021524 .

Bem DJ, Utts J, Johnson WO (2011). “Must psychologists change the way they analyze their data?” Journal of Personality and Social Psychology, 101(4), 716–719. doi:10.1037/a0024777 .