1159 effect sizes from a meta-analysis of beauty and professional success by Havránková et al. (2025)

The data set contains effect sizes (percent increase in earnings), standard errors, study identifiers, sample sizes, and the type of customer contact (no, some, or direct). The meta-analysis examined the relationship between perceived beauty and professional success (Havránková et al. 2025) .

Usage

Havrankova2025

Format

A data.frame with 5 columns and 1159 observations:

y

Effect size: percent increase in earnings.

se

Standard error of y.

facing_customer

Type of customer contact.

study_id

Study identifier.

N

Sample size.

References

Havránková Z, Havránek T, Bortnikova K, Bartoš F (2025). “Meta-analysis of field studies on beauty and professional success.” Preprint available at https://meta-analysis.cz/beauty/beauty.pdf.