412 effect sizes from a meta-analysis of secondary benefits of family-based treatments by Johnides et al. (2025)

The data set contains Cohen's d effect sizes, standard errors, and study labels from a meta-analysis investigating the extent to which family-based treatments for children with mental health, physical health, and developmental disorders provide secondary benefits to the children's siblings and caregivers (Johnides et al. 2025) .

Usage

Johnides2025

Format

A data.frame with 3 columns and 412 observations:

study

Study label.

d

Cohen's d effect size.

se

Standard error of d.

References

Johnides BD, Borduin CM, Sheerin KM, Kuppens S (2025). “Secondary benefits of family member participation in treatments for childhood disorders: A multilevel meta-analytic review.” Psychological Bulletin, 151(1), 1–32. doi:10.1037/bul0000462 .