55 effect sizes from Many Labs 2 replication studies of Tversky and Kahneman (1981) framing effects

The data set contains standardized mean differences (y) and standard errors (se) from 55 replication studies of the framing effect originally described by Tversky and Kahneman (1981) . These studies were part of the Many Labs 2 project examining variation in replicability across samples and settings Klein et al. (2018) .

Usage

ManyLabs16

Format

A data.frame with 2 columns and 55 observations:

y

Standardized mean difference.

se

Standard error of y.

References

Klein RA, Vianello M, Hasselman F, Adams BG, Adams Jr RB, Alper S, Aveyard M, Axt JR, Babalola MT, Bahník Š, others (2018). “Many Labs 2: Investigating variation in replicability across samples and settings.” Advances in Methods and Practices in Psychological Science, 1(4), 443–490. doi:10.1177/2515245918810225 .

Tversky A, Kahneman D (1981). “The framing of decisions and the psychology of choice.” Science, 211(4481), 453–458. doi:10.1126/science.7455683 .