Create Histogram of Z-Statistics

Plots a histogram of observed z-values for a fitted zcurve_RoBMA object, with options to customize the plotting range, bin width, and display of significance thresholds.

Usage

# S3 method for class 'zcurve_RoBMA'
hist(
  x,
  plot_type = "base",
  from = -6,
  to = 6,
  by = 0.5,
  length.out = NULL,
  plot_thresholds = TRUE,
  ...
)

Arguments

x

A zcurve_RoBMA object containing the fitted model.

plot_type

whether to use a base plot "base" or ggplot2 "ggplot" for plotting. Defaults to "base".

from

Lower bound of the z-value range for plotting. Defaults to -6.

to

Upper bound of the z-value range for plotting. Defaults to 6.

by

Numeric value specifying the bin width for the histogram. Defaults to 0.5.

length.out

Optional integer specifying the number of bins. If NULL, determined by by. Defaults to NULL.

plot_thresholds

Logical; should significance thresholds be displayed on the plot? Defaults to TRUE.

...

Additional arguments passed to the underlying plotting functions.

Value

Returns NULL if plot_type = "base", or a ggplot2 object if plot_type = "ggplot2".

See also