Summarize brma_samples Object

Creates and returns a summary table of posterior samples using BayesTools::ensemble_estimates_table.

Usage

# S3 method for class 'brma_samples'
summary(object, probs = NULL, ...)

Arguments

object

a brma_samples object

probs

quantiles for credible intervals. If NULL, uses the default stored in the object (typically c(.025, .975))

...

additional arguments passed to BayesTools::ensemble_estimates_table

Value

A BayesTools_table object containing the summary statistics.