Anderson2010 {RoBMA}	R Documentation

27 experimental studies from from Anderson et al. (2010) that meet the best practice criteria

Description

The dataset contains correlational coefficients, sample sizes, and labels for 27 experimental studies focusing on the effect of violent videogames on aggressive behavior. The full original data can found at https://github.com/Joe-Hilgard/Anderson-meta.

Usage


Anderson2010

Format

A data.frame with 3 columns and 23 observations.

References

Anderson CA, Shibuya A, Ihori N, Swing EL, Bushman BJ, Sakamoto A, Rothstein HR, Saleem M (2010). “Violent video game effects on aggression, empathy, and prosocial behavior in Eastern and Western countries: A meta-analytic review.” Psychological bulletin, 136(2), 151. Publisher: American Psychological Association.

[Package RoBMA version 1.2.0.902 Index]