RoBMA: Robust Bayesian meta-analysis

RoBMA: Bayesian model-averaged meta-analysis with adjustments for publication bias and ability to specify informed prior distributions and draw inference with inclusion Bayes factors.

User guide

See Bartoš et al. (2021) , Maier et al. (in press) , and Bartoš et al. (2020) for details regarding the RoBMA methodology.

More details regarding customization of the model ensembles are provided in the Reproducing BMA and Fitting custom meta-analytic ensembles vignettes. Please, use the "Issues" section in the GitHub repository to ask any further questions.

References

Bartoš F, Maier M, Wagenmakers E (2020). “Adjusting for Publication Bias in JASP — Selection Models and Robust Bayesian Meta-Analysis.” preprint at https://doi.org/10.31234/osf.io/75bqn, https://doi.org/10.31234/osf.io/75bqn.

Bartoš F, Maier M, Wagenmakers E, Doucouliagos H, Stanley TD (2021). “Need to choose: Robust Bayesian meta-analysis with competing publication bias adjustment methods.” preprint at https://doi.org/10.31234/osf.io/kvsp7, https://doi.org/10.31234/osf.io/kvsp7.

Maier M, Bartoš F, Wagenmakers E (in press). “Robust Bayesian Meta-Analysis: Addressing Publication Bias with Model-Averaging.” Psychological Methods. https://doi.org/10.31234/osf.io/u4cns.

See also

Author

František Bartoš f.bartos96@gmail.com