RoBMA: Robust Bayesian meta-analysis

RoBMA: Bayesian model-averaged meta-analysis with adjustments for publication bias and ability to specify informed prior distributions and draw inference with inclusion Bayes factors.

User guide

See bartos2021no;textualRoBMA, maier2020robust;textualRoBMA, and bartos2020adjusting;textualRoBMA for details regarding the RoBMA methodology.

More details regarding customization of the model ensembles are provided in the Reproducing BMA, BMA in Medicine, and Fitting Custom Meta-Analytic Ensembles vignettes. Please, use the "Issues" section in the GitHub repository to ask any further questions.

References

See also

Author

František Bartoš f.bartos96@gmail.com