RoBMA-package {RoBMA}	R Documentation

RoBMA: Robust Bayesian Meta-Analyses

Description

The package estimates an ensemble of Bayesian model-averaged meta-analytical models (assuming either presence or absence of the effect, heterogeneity, and publication bias). The ensemble uses Bayes Factors to test for presence of absence of the individual components and model-averages parameter estimates based on posterior model probabilities. The user can define a wide range of non-informative or informative priors for the effect size, heterogeneity, and weight functions. The package provides convenient functions for summary, visualizations, and fit diagnostics.

Author(s)

Maintainer: František Bartoš f.bartos96@gmail.com (ORCID)

Authors:

Maximilian Maier maximilianmaier0401@gmail.com (ORCID)

Other contributors:

Eric-Jan Wagenmakers EJ.Wagenmakers@gmail.com (ORCID) [thesis advisor]

Joris Goosen j.c.goosen@uva.nl [contributor]

[Package RoBMA version 0.0.0.9000 Index]