RoBMA-package {RoBMA}	R Documentation

RoBMA: Robust Bayesian Meta-Analyses

Description

A framework for estimating ensembles of meta-analytic models (assuming either presence or absence of the effect, heterogeneity, and publication bias) and using Bayesian model averaging to combine them. The ensembles use Bayes factors to test for the presence or absence of the individual components (e.g., effect vs. no effect) and model-averages parameter estimates based on posterior model probabilities (Maier, Bartoš & Wagenmakers, 2020, <doi:10.31234/osf.io/u4cns>). The user can define a wide range of non-informative or informative priors for the effect size, heterogeneity, and weight functions. The package provides convenient functions for summary, visualizations, and fit diagnostics.

Author(s)

Maintainer: František Bartoš f.bartos96@gmail.com (ORCID)

Authors:

Maximilian Maier maximilianmaier0401@gmail.com (ORCID)

Other contributors:

Eric-Jan Wagenmakers (ORCID) [thesis advisor]

Joris Goosen [contributor]

See Also

Useful links:

https://fbartos.github.io/RoBMA/

Report bugs at https://github.com/FBartos/RoBMA/issues

[Package RoBMA version 1.0.5 Index]