Options for the RoBMA package

A placeholder object and functions for the RoBMA package. (adapted from the runjags R package).

Usage

RoBMA.options(...)

RoBMA.get_option(name)

Arguments

...

named option(s) to change - for a list of available options, see details below.

name

the name of the option to get the current value of - for a list of available options, see details below.

Value

The current value of all available RoBMA options (after applying any changes specified) is returned invisibly as a named list.