Prints summary of "RoBMA" ensemble implied by the specified priors

check_setup prints summary of "RoBMA" ensemble implied by the specified prior distributions. It is useful for checking the ensemble configuration prior to fitting all of the models.

check_setup(
  priors_mu = prior(distribution = "normal", parameters = list(mean = 0, sd = 1)),
  priors_tau = prior(distribution = "invgamma", parameters = list(shape = 1, scale =
    0.15)),
  priors_omega = list(prior(distribution = "two.sided", parameters = list(alpha = c(1,
    1), steps = c(0.05)), prior_odds = 1/2), prior(distribution = "two.sided", parameters
    = list(alpha = c(1, 1, 1), steps = c(0.05, 0.1)), prior_odds = 1/2)),
  priors_mu_null = prior(distribution = "point", parameters = list(location = 0)),
  priors_tau_null = prior(distribution = "point", parameters = list(location = 0)),
  priors_omega_null = prior(distribution = "point", parameters = list(location = 1)),
  models = FALSE,
  silent = FALSE
)

Arguments

priors_mu
list of prior distributions for the mu parameter that will be treated as belonging to the alternative hypothesis. Defaults to prior(distribution = "normal", parameters = list(mean = 0, sd = 1)).

priors_tau
list of prior distributions for the tau parameter that will be treated as belonging to the alternative hypothesis. Defaults to prior(distribution = "invgamma", parameters = list(shape = 1, scale = .15)).

priors_omega
list of prior weight functions for the omega parameter that will be treated as belonging to the alternative hypothesis. Defaults to list( prior(distribution = "two.sided", parameters = list(alpha = c(1, 1), steps = c(.05)), prior_odds = 1/2), prior(distribution = "two.sided", parameters = list(alpha = c(1, 1, 1), steps = c(.05, .10)), prior_odds = 1/2) ).

priors_mu_null
list of prior distributions for the mu parameter that will be treated as belonging to the null hypothesis. Defaults to point distribution with location at 0 ( prior(distribution = "point", parameters = list(location = 0))).

priors_tau_null
list of prior distributions for the tau parameter that will be treated as belonging to the null hypothesis. Defaults to point distribution with location at 0 ( prior(distribution = "point", parameters = list(location = 0))).

priors_omega_null
list of prior weight functions for the omega parameter that will be treated as belonging to the null hypothesis. Defaults to point distribution with location at 1 ( prior(distribution = "point", parameters = list(location = 0))).

models
should the models' details be printed.

silent
do not print the results.

See also